+++ Database/dbo/Tables/UM_Department.sql
@@ -5,10 +5,12 @@
     [Description]          NVARCHAR (MAX) NULL,
     [ResponsibilityCenter] NVARCHAR (MAX) NULL,
     [IsActive]             BIT            NOT NULL,
-    [CreatedByUserId]      BIGINT         NOT NULL,
+    [CreatedByUserId]      INT            NOT NULL,
     [CreatedDate]          DATETIME       NOT NULL,
     [BureauId]             INT            NULL,
     CONSTRAINT [PK_UM_Department] PRIMARY KEY CLUSTERED ([DepartmentId] ASC),
     CONSTRAINT [FK_UM_Department_UM_Bureau] FOREIGN KEY ([BureauId]) REFERENCES [dbo].[UM_Bureau] ([BureauId])
 );
+
+
